requirements.txt
django
djangorestframework
drf_yasg
djoser==0.4.1
psycopg2-binary